fix hist.py, mini_maxwell.py, status_light.py for python2/python3 compatible

+++ jsk_rqt_plugins/src/jsk_rqt_plugins/mini_maxwell.py
@@ -67,7 +67,7 @@ class DRCEnvironmentViewerWidget(QWidget):
             Time, self.nextWhiteoutTimeCallback)
         self._update_plot_timer = QTimer(self)
         self._update_plot_timer.timeout.connect(self.redraw)
-        self._update_plot_timer.start(1000 / 15)
+        self._update_plot_timer.start(int(1000 / 15))
 
     def isDisabledCallback(self, msg):
         with self.lock:
